chara_params
